src/requirements.txt
# 基础依赖
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
PyQt5==5.15.10
python-docx==0.8.11
openpyxl==3.1.2
markdown==3.5.1
python-pptx==0.6.21
beautifulsoup4==4.12.2

# AI模型相关
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99
accelerate==0.27.2 